Revert "mach-midas: Increase USB charging rate"
Majority of SDP ports indeed adhere to the 500mA limit, setting the
charging current by default high triggers soft regulation to kick in
and decreases in 100mA steps until stable power is reached. The 25mA
margin provided by the default 475 is enough for stable power,
otherwise softregulation sets it back to 400mA (300mA real).

Revert everything to default and let users choose it with the charging
interface which comes next.

This reverts commit 22b32d44ce003c3e126fbc3fab9615b4c5c3980a.