touchwake: ignore_once sysfs entry
when a virtual power key event is triggered, touchwake has no way of
knowing that wasn't a timeout, so it activates. this means there is
no easy way to enter suspend from userspace without always tripping
touchwake. so this sysfs entry will block touckwake from running the
next, and only next, time it is called if it is set to 1. then your
script or whatever can normally call for suspend and touchwake will
ignore it that once.