mms152_ts: plasma touchbooster v1.1
reworked the way the touchbooster operates. previously, on each finger-up event
the booster would schedule work to deactivate itself, regardless of the fact that
there may be another finger down already. so even if you had another finger down
within the booster duration, it would still insist on turning itself off and that
new finger would have to turn the booster on again.

- new behavior "bounces" the most recent tap, like a dolphin with a ball, so the booster
  now never turns off unless all fingers have been up for the specified duration.

- also, inherently this allows for better relaxed-mode operation, as relaxed-mode will
  now take effect during the duration period mentioned in the sentence above.

- turned logging on for now in anticipation of any bugs

- lowered the sanitization of the typingbooster_mintaps sysfs setting to allow for
  values as low as 1.