cpufreq/zzmoove: check profile integrity only if value changed
zzmoove automatically sets the profile to 0 (custom) when any tuneable
is written to, regardless of whether or not the value was actually changed.
this causes problems when some utilities routinely rewrite values,
as the profile always reverts to 0 even if no values were actually changed.

this commit changes that behavior, so profiles only "break" when a protected
tuneable is actually changed.