cpufreq/zzmoove: add profiles_sticky_mode
in some situations it may be useful to tweak profiles without
profile_number reverting to 0. granted, this carries the risk
of not being able to tell if the profile is fully applied or not.
however, for my usage i'd rather it not always jump to the ambiguous
"custom" profile.

- profiles_sticky_mode
   0 = normal (profile "breaks" when any tuneable changed)
   1 = sticky mode (profile stays intact, even if a tuneable changes)

in the future, a sysfs can be made to verify the integrity of the
currently set profile.