cpufreq/zzmoove: keep dynamic tuneables constant, typingbooster bugfix
dbs_tuners_ins.sampling_down_factor and dbs_tuners_ins.sampling_down_max_mom
dynamically change, which in turn makes some utilities go crazy.

this commit allows the tuneables to remain constant, while a duplicate variable
is deals with the governor's dynamic values.

also includes a bugfix for the typingbooster, in the form of the added _saved
variables in the up_threshold and hotplug_min_limit portion of set_profile().